Conflicts:
	src/main/java/kr/co/dong/util/GoogleApiUtil.java
	src/main/webapp/WEB-INF/views/include/left_nav.jsp

+++ src/main/java/kr/co/dong/MemberController.java
@@ -27,6 +27,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.scribejava.core.model.OAuth2AccessToken;
+import com.mysql.cj.Session;
 
 import kr.co.dong.member.MemberDTO;
 import kr.co.dong.member.MemberService;
@@ -244,71 +245,84 @@ public class MemberController {
 // 관리자 회원관리 전체 리스트
    @RequestMapping("/members")
    public String members(@RequestParam(value="currentPage", required=false, defaultValue="1") int currentPage,
-                    Model model){
-      
-      int limit = 15;   // 페이지당 목록 수
-      int offset = (currentPage -1) * limit;
-      
-      // 검색조건 Map int형 따로 분리 object에서 안들어감
-      Map<String,Object> params = new HashMap<>();
-      params.put("limit", Integer.valueOf(limit));
-      params.put("offset", Integer.valueOf(offset));
-      
-      
-      int totalCount = memberService.memberCount(); // 총 회원 수
-      List<MemberDTO> list = memberService.allList(params); // 전체회원 목록
-      int searchMembersCount = memberService.searchMembersCount(params); // 검색 회원 수
-      
-      int totalPages = totalCount / limit;   // 정수 나눗셈 = 자동 소수점 버림
-       if(totalCount % limit != 0) {         // 나머지가 있으면 한 페이지 추가
-           totalPages += 1;
-       }
-       
-      if(currentPage > totalPages) currentPage = totalPages;
-      if(currentPage < 1) currentPage = 1; 
-      
+                    Model model, HttpSession session){
+	   String role = (String) session.getAttribute("role");
+      if("ADMIN".equals(role)) {
+    	  int limit = 15;   // 페이지당 목록 수
+          int offset = (currentPage -1) * limit;
+          
+          // 검색조건 Map int형 따로 분리 object에서 안들어감
+          Map<String,Object> params = new HashMap<>();
+          params.put("limit", Integer.valueOf(limit));
+          params.put("offset", Integer.valueOf(offset));
+          
+          
+          int totalCount = memberService.memberCount(); // 총 회원 수
+          List<MemberDTO> list = memberService.allList(params); // 전체회원 목록
+          int searchMembersCount = memberService.searchMembersCount(params); // 검색 회원 수
+          
+          int totalPages = totalCount / limit;   // 정수 나눗셈 = 자동 소수점 버림
+           if(totalCount % limit != 0) {         // 나머지가 있으면 한 페이지 추가
+               totalPages += 1;
+           }
+           
+          if(currentPage > totalPages) currentPage = totalPages;
+          if(currentPage < 1) currentPage = 1; 
+          
 
-      model.addAttribute("list", list);
-      model.addAttribute("totalCount", totalCount);
-      model.addAttribute("searchCount", totalCount);   // 검색카운트
-      model.addAttribute("currentPage", currentPage);
-      model.addAttribute("totalPages", totalPages);
-      model.addAttribute("limit", limit);
-      
-      return "admin/members";
+          model.addAttribute("list", list);
+          model.addAttribute("totalCount", totalCount);
+          model.addAttribute("searchCount", totalCount);   // 검색카운트
+          model.addAttribute("currentPage", currentPage);
+          model.addAttribute("totalPages", totalPages);
+          model.addAttribute("limit", limit);
+          
+          return "admin/members";
+    	  
+      }else
+   		   session.invalidate(); // 세션 초기화
+	   return "error/405?msg=invalidAccess";
+     
    }
 
    // 관리자 회원검색 리스트
    @RequestMapping("/members_search")
    public String searchMembers(@RequestParam Map<String, Object> params, // 모든 파라미터를 한
                         @RequestParam(value="currentPage", required=false, defaultValue="1") int currentPage,
-                        Model model) {
-      
-      //2. 검색 필터조건 Map 에 넣어 서비스로 보내기
-      int limit = 15;   // 페이지당 목록 수
-      int offset = (currentPage -1) * limit; 
-      
-      // 검색조건 Map int형 따로 분리 object에서 안들어감
-      params.put("limit", Integer.valueOf(limit));
-      params.put("offset", Integer.valueOf(offset));
-      
-      
-//      List<MemberDTO> list= memberService.searchMembers(searchType, searchValue); // 이전꺼
-      List<MemberDTO> list= memberService.searchMembers(params);   //검색 회원 목록
+                        Model model, HttpSession session) {
+	   String role = (String) session.getAttribute("role");
+	   if("ADMIN".equals(role)) {
+		   //2. 검색 필터조건 Map 에 넣어 서비스로 보내기
+		      int limit = 15;   // 페이지당 목록 수
+		      int offset = (currentPage -1) * limit; 
+		      
+		      // 검색조건 Map int형 따로 분리 object에서 안들어감
+		      params.put("limit", Integer.valueOf(limit));
+		      params.put("offset", Integer.valueOf(offset));
+		      
+		      
+//		      List<MemberDTO> list= memberService.searchMembers(searchType, searchValue); // 이전꺼
+		      List<MemberDTO> list= memberService.searchMembers(params);   //검색 회원 목록
 
-      int totalCount = memberService.memberCount();   // 총 회원수
-      int searchMembersCount = memberService.searchMembersCount(params); // 검색 회원 수
-      int totalPage = (int) Math.ceil((double) searchMembersCount / limit);   // 총페이지수 계산
+		      int totalCount = memberService.memberCount();   // 총 회원수
+		      int searchMembersCount = memberService.searchMembersCount(params); // 검색 회원 수
+		      int totalPage = (int) Math.ceil((double) searchMembersCount / limit);   // 총페이지수 계산
 
-      model.addAttribute("list", list);
-      model.addAttribute("totalCount", totalCount);
-      model.addAttribute("searchCount", searchMembersCount);
-      model.addAttribute("params", params);
-      model.addAttribute("currentPage", currentPage);
-      model.addAttribute("totalPage", totalPage);
+		      model.addAttribute("list", list);
+		      model.addAttribute("totalCount", totalCount);
+		      model.addAttribute("searchCount", searchMembersCount);
+		      model.addAttribute("params", params);
+		      model.addAttribute("currentPage", currentPage);
+		      model.addAttribute("totalPage", totalPage);
 
-      return "admin/members";
+		      return "admin/members";
 
+		   
+	   }else
+		   session.invalidate(); // 세션 초기화
+	   return "error/405?msg=invalidAccess";
+	  
+    
    }
 
    // 회원정보 수정 페이지 내용 삽입
@@ -326,10 +340,16 @@ public class MemberController {
 
    // 관리자 회원정보 수정 페이지 내용 삽입
    @PostMapping(value = "/adminupdate")
-   public String adminupdate(@RequestParam("id") String id, Model model) {
-       MemberDTO list = memberService.selectone(id);
-        model.addAttribute("user", list);  
-        return "admin/adminupdate";  
+   public String adminupdate(@RequestParam("id") String id, Model model, HttpSession session) {
+	   String role = (String) session.getAttribute("role");
+	   if("ADMIN".equals(role)) {
+		   MemberDTO list = memberService.selectone(id);
+	        model.addAttribute("user", list);  
+	        return "admin/adminupdate";  
+	   }else
+		   session.invalidate(); // 세션 초기화
+	   return "error/405?msg=invalidAccess";
+      
    }
 
    // 회원 정보 수정
@@ -341,15 +361,27 @@ public class MemberController {
    
    // 관리자 회원 정보 수정
    @PostMapping(value = "/adminupdate_ok")
-   public String adminupdate(@ModelAttribute MemberDTO update) {
-      memberService.adminupdate(update);
-        return "redirect:/member/members";  
+   public String adminupdate(@ModelAttribute MemberDTO update, HttpSession session) {
+	   String role = (String) session.getAttribute("role");
+	   if("ADMIN".equals(role)) {
+		   memberService.adminupdate(update);
+	        return "redirect:/member/members";  
+	   }else
+		   session.invalidate(); // 세션 초기화
+	   return "error/405?msg=invalidAccess";
+      
    }
    // 관리자 삭제
    @PostMapping(value = "/deleteadmin")
-   public String deleteadmin(@RequestParam("id") String id) {
-      memberService.deleteadmin(id);
-      return "redirect:/member/members";  
+   public String deleteadmin(@RequestParam("id") String id, HttpSession session) {
+	   String role = (String) session.getAttribute("role");
+	   if("ADMIN".equals(role)) {
+		   memberService.deleteadmin(id);
+		      return "redirect:/member/members";  
+	   }else
+		   session.invalidate(); // 세션 초기화
+	   return "error/405?msg=invalidAccess";
+     
    }
    @PostMapping("/change-password")
    @ResponseBody
@@ -473,10 +505,16 @@ public class MemberController {
    
    // 관리자 회원상세조회 체이지
    @GetMapping("/detail")
-   public String detail(@RequestParam("id") String id, Model model) {
-      MemberDTO user = memberService.selectone(id);
-      model.addAttribute("user", user);
-      return "admin/detail";
+   public String detail(@RequestParam("id") String id, Model model, HttpSession session) {
+	   String role = (String) session.getAttribute("role");
+	   if("ADMIN".equals(role)) {
+		   MemberDTO user = memberService.selectone(id);
+		      model.addAttribute("user", user);
+		      return "admin/detail";
+	   }else
+		   session.invalidate(); // 세션 초기화
+	   return "error/405?msg=invalidAccess";
+     
       
    }
    
@@ -539,10 +577,6 @@ public class MemberController {
 		}
 	}
 	
-	// 통계페이지 동
-	@GetMapping("/stats")
-	public String stats() {
-		return "admin/stats"; 
-	}
+
 	
 }
